Changes OK

+++ src/main/java/ejercicios/dao/PiezaDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ejercicios.dto.Pieza;
 
-public interface PiezaDAO extends JpaRepository<Pieza, Long>{
+public interface PiezaDAO extends JpaRepository<Pieza, Integer>{
 
 }
